Refactor chat files and create MessageAdapter
